refactor the adaptors to clean the authentication folder
