+++ pkg/authentication/adapters/graph/schema.graphqls
