+++ pkg/authentication/adapters/pb/auth.proto
