+++ pkg/authentication/adapters/resolver.go
@@ -1,12 +1,15 @@
-package authentication
+package adapters
 
-import "log"
+import (
+	"github.com/Hamifthi/authentication_microservice/pkg/authentication"
+	"log"
+)
 
 // This file will not be regenerated automatically.
 //
 // It serves as dependency injection for your app, add any dependencies you require here.
 
 type Resolver struct {
-	AuthService *authenticationService
+	AuthService *authentication.AuthenticationService
 	Logger      *log.Logger
 }
